3
Renamed each file to represent it's target OS and removed the version
number from the filename.
I added some comments to make the project easier to follow.
I still have a lot of code to fix.  Did I really not initialize my
variables at declaration?  Was that normal in the 90's, or was I just a
bad programmer?
I plan on moving file opening to main() for scope reasons, and moving
code common to both programs into a header file.